Simplify stream http adapter
