+++ src/Widop/HttpAdapter/StreamHttpAdapter.php
@@ -15,8 +15,9 @@ namespace Widop\HttpAdapter;
  * Stream Http adapter.
  *
  * @author Geoffrey Brier <geoffrey.brier@gmail.com>
+ * @author GeLo <geloen.eric@gmail.com>
  */
-class StreamHttpAdapter implements HttpAdapterInterface
+class StreamHttpAdapter extends AbstractHttpAdapter
 {
     /**
      * {@inheritdoc}
@@ -40,14 +41,14 @@ class StreamHttpAdapter implements HttpAdapterInterface
      * @param string   $url     An url.
      * @param resource $context A resource (created from stream_context_create).
      *
-     * @throws \Widop\HttpAdapterBundle\Exception\HttpAdapterException On url opening/fetching error.
+     * @throws \Widop\HttpAdapterBundle\Exception\HttpAdapterException If an error occured.
      *
      * @return string The response content.
      */
     protected function execute($url, $context)
     {
         if (($fp = @fopen($this->fixUrl($url), 'rb', false, $context)) === false) {
-            throw HttpAdapterException::cannotOpenUrl($url, $this->getName(), print_r(error_get_last(), true));
+            throw HttpAdapterException::cannotFetchUrl($url, $this->getName(), print_r(error_get_last(), true));
         }
 
         $content = stream_get_contents($fp);
@@ -66,115 +67,34 @@ class StreamHttpAdapter implements HttpAdapterInterface
     }
 
     /**
-     * Fixes the URL (adds http:// if not set).
-     *
-     * @param string $url An URL.
-     *
-     * @return string A valid URL.
-     */
-    protected function fixUrl($url)
-    {
-        if (strpos($url, 'http://') !== 0 && strpos($url, 'https://') !== 0) {
-            return 'http://'.$url;
-        }
-
-        return $url;
-    }
-
-    /**
      * Creates the stream context.
      *
-     * @param string $method  The HTTP method (eg: GET, POST).
-     * @param array  $headers An array of headers.
-     * @param string $content The content.
+     * @param string       $method  The HTTP method.
+     * @param array        $headers The headers.
+     * @param string|array $content The content.
      *
      * @return resource A stream context resource.
      */
     protected function createStreamContext($method, array $headers, $content = '')
     {
-        $rationalizedHeaders = array();
         $contextOptions = array('http' => array('method' => $method));
 
-        // Rationalizes headers as an associative array
         if (!empty($headers)) {
-            foreach ($headers as $key => $value) {
-                if (is_int($key)) {
-                    list ($key, $value) = $this->extractHeaderKeyAndValue($value);
-                }
+            $contextOptions['http']['header'] = '';
 
-                $rationalizedHeaders[$key] = trim($value);
+            foreach ($this->fixHeaders($headers) as $header) {
+                $contextOptions['http']['header'] .= sprintf("%s\r\n", $header);
             }
         }
 
-        // Sets POST data
         if ($method === 'POST') {
             if (is_array($content)) {
                 $content = http_build_query($content);
             }
 
             $contextOptions['http']['content'] = $content;
-
-            if (!$this->headerKeyMatches($rationalizedHeaders, 'Content-Length')) {
-                $rationalizedHeaders['Content-Length'] = strlen($content);
-            }
-            if (!$this->headerKeyMatches($rationalizedHeaders, 'Content-type')) {
-                $rationalizedHeaders['Content-type'] = 'application/x-www-form-urlencoded';
-            }
-        }
-
-        // Set headers
-        if (!empty($rationalizedHeaders)) {
-            $contextOptions['http']['header'] = '';
-            foreach ($rationalizedHeaders as $hKey => $hValue) {
-                $contextOptions['http']['header'] .= "$hKey:$hValue\r\n";
-            }
         }
 
         return stream_context_create($contextOptions);
-    }
-
-    /**
-     * Extracts an header key ('Content-Length') and an header value ('42') from an
-     * header line ('Content-Length: 42').
-     *
-     * @param string $header The header line.
-     *
-     * @throws \Widop\HttpAdapterBundle\Exception\HttpAdapterException On invalid header.
-     *
-     * @return array An array representing a header (0: Content-Length, 1: '42').
-     */
-    protected function extractHeaderKeyAndValue($header)
-    {
-        if (($pos = strpos($header, ':')) === false) {
-            throw HttpAdapterException::invalidHeader($header, 'Missing ":"');
-        }
-
-        $key = substr($header, 0, $pos);
-        $value = trim(substr($header, $pos + 1));
-
-        if (empty($key) || empty($value)) {
-            throw HttpAdapterException::invalidHeader($header, 'Empty key or value');
-        }
-
-        return array($key, $value);
-    }
-
-    /**
-     * Checks if $headerKey is in the associative array of headers (case insensitive).
-     *
-     * @param array  $headers   An associative array of headers ('Content-Type' => 'html/css').
-     * @param string $headerKey The header key to look for.
-     *
-     * @return boolean True when the key was found.
-     */
-    protected function headerKeyMatches(array $headers, $headerKey)
-    {
-        foreach (array_keys($headers) as $hKey) {
-            if (strcasecmp($hKey, $headerKey) === 0) {
-                return true;
-            }
-        }
-
-        return false;
     }
 }
