+++ tests/Widop/Tests/HttpAdapter/CurlHttpAdapterTest.php
@@ -32,13 +32,4 @@ class CurlHttpAdapterTest extends AbstractHttpAdapterTest
     {
         $this->assertSame('curl', $this->httpAdapter->getName());
     }
-
-    /**
-     * @expectedException \Widop\HttpAdapter\HttpAdapterException
-     */
-    public function testExecuteThrowsInvalidCallbackException()
-    {
-        $this->getHttpAdapterReflectionMethod('execute')
-            ->invokeArgs($this->httpAdapter, array('foo', array(), '', 'tata'));
-    }
 }
