+++ tests/Widop/Tests/HttpAdapter/StreamHttpAdapterTest.php
@@ -32,87 +32,4 @@ class StreamHttpAdapterTest extends AbstractHttpAdapterTest
     {
         $this->assertSame('stream', $this->httpAdapter->getName());
     }
-
-    /**
-     * @expectedException \Widop\HttpAdapter\HttpAdapterException
-     */
-    public function testExtractMethodWithEmptyHeaderThrowsException()
-    {
-        $this->getHttpAdapterReflectionMethod('extractHeaderKeyAndValue')
-            ->invokeArgs($this->httpAdapter, array(''));
-    }
-
-    /**
-     * @expectedException \Widop\HttpAdapter\HttpAdapterException
-     */
-    public function testExtractMethodWithInvalidHeaderThrowsException()
-    {
-        $this->getHttpAdapterReflectionMethod('extractHeaderKeyAndValue')
-            ->invokeArgs($this->httpAdapter, array('foo'));
-    }
-
-    /**
-     * @expectedException \Widop\HttpAdapter\HttpAdapterException
-     */
-    public function testExtractMethodWithEmptyHeaderKeyThrowsException()
-    {
-        $this->getHttpAdapterReflectionMethod('extractHeaderKeyAndValue')
-            ->invokeArgs($this->httpAdapter, array(': 42\r\n'));
-    }
-
-    /**
-     * @expectedException \Widop\HttpAdapter\HttpAdapterException
-     */
-    public function testExtractMethodWithEmptyHeaderValueThrowsException()
-    {
-        $this->getHttpAdapterReflectionMethod('extractHeaderKeyAndValue')
-            ->invokeArgs($this->httpAdapter, array('Content-Length:'));
-    }
-
-    public function testExtractMethodReturnsCorrectValues()
-    {
-        list ($hKey, $hValue) = $this->getHttpAdapterReflectionMethod('extractHeaderKeyAndValue')
-            ->invokeArgs($this->httpAdapter, array("Content-Length: 42\r\n"));
-
-        $this->assertEquals('Content-Length', $hKey);
-        $this->assertEquals('42', $hValue);
-    }
-
-    public function testHeaderKeyMatchesIsCaseInsensitive()
-    {
-        $doMatch = $this->getHttpAdapterReflectionMethod('headerKeyMatches')
-            ->invokeArgs($this->httpAdapter, array(array('Content-Length' => '42'), 'CoNtenT-LenGth'));
-
-        $this->assertTrue($doMatch);
-    }
-
-    public function testHeaderKeyMatchesReturnsFalse()
-    {
-        $doMatch = $this->getHttpAdapterReflectionMethod('headerKeyMatches')
-            ->invokeArgs($this->httpAdapter, array(array('Content-type' => 'html/css'), 'CoNtenT-LenGth'));
-
-        $this->assertFalse($doMatch);
-
-    }
-
-    public function testFixUrlAddsHttpSchemeIfNotPresent()
-    {
-        $url = $this->getHttpAdapterReflectionMethod('fixUrl')
-            ->invokeArgs($this->httpAdapter, array('www.google.fr'));
-
-        $this->assertEquals('http://www.google.fr', $url);
-    }
-
-    public function testFixUrlDoesNotModifyUrlIfSchemePresent()
-    {
-        $httpUrl = $this->getHttpAdapterReflectionMethod('fixUrl')
-            ->invokeArgs($this->httpAdapter, array('http://www.google.fr'));
-
-        $this->assertEquals('http://www.google.fr', $httpUrl);
-
-        $httpsUrl = $this->getHttpAdapterReflectionMethod('fixUrl')
-            ->invokeArgs($this->httpAdapter, array('https://www.google.fr'));
-
-        $this->assertEquals('https://www.google.fr', $httpsUrl);
-    }
 }
